refactor: move ErrorBoundary to components folder
